refactor: createRacingGame에서 사용자 입력값 검증 로직 Controller로 분리해 도메인 로직과 UI 로직 분리
